Experience, Spawner
- Created a form of Experience Distribution when an agent dies
- Reworked Spawner
-- Separated Agent Creation from spawning
-- Now I can create an agent based on an Agent Description which I can expand down the line
- Fixed some pathfinding bugs

+++ Assets/Scripts/Gameplay/Agent/Actions/Charge.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using System.Collections.Generic;
 public class Charge : AgentAction
 {
-    public Charge(int startingLevel, float expLevelUp) : base("Charge", startingLevel, 99, expLevelUp)
+    public Charge(int startingLevel, int expLevelUp) : base("Charge", startingLevel, 99, expLevelUp)
     {
         baseDamage = 1;
         Range = 5;
